move wallet to its own db
