+++ src/Features/Blockcore.Features.NodeHost/UI/Pages/_Host.cshtml
@@ -1,7 +1,6 @@
 ﻿@page "/"
 @inject NBitcoin.Network Network
 @namespace Blockcore.Features.NodeHost.UI.Pages
-@inject NBitcoin.Network Network
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, BlazorModal
 @{
