+++ src/Features/Blockcore.Features.WalletWatchOnly/API/Controllers/WatchOnlyWalleRPCController.cs
@@ -8,6 +8,7 @@ using Blockcore.Controllers.Models;
 using Blockcore.Features.RPC;
 using Blockcore.Features.RPC.Exceptions;
 using Blockcore.Features.Wallet;
+using Blockcore.Features.Wallet.Database;
 using Blockcore.Features.Wallet.Interfaces;
 using Blockcore.Features.Wallet.Types;
 using Blockcore.Features.WalletWatchOnly.Api.Models;
