+++ src/Tests/Blockcore.Features.Miner.Tests/PosMintingTest.cs
@@ -16,6 +16,7 @@ using Blockcore.Features.MemoryPool.Interfaces;
 using Blockcore.Features.Miner.Api.Models;
 using Blockcore.Features.Miner.Staking;
 using Blockcore.Features.Wallet;
+using Blockcore.Features.Wallet.Database;
 using Blockcore.Features.Wallet.Interfaces;
 using Blockcore.Features.Wallet.Tests;
 using Blockcore.Features.Wallet.Types;
