+++ src/Tests/Blockcore.Features.Wallet.Tests/HdAddressTest.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Blockcore.Features.Wallet.Database;
 using Blockcore.Features.Wallet.Types;
 using NBitcoin;
 using Xunit;
