+++ src/Tests/Blockcore.Features.Wallet.Tests/WalletControllerTest.cs
@@ -8,6 +8,7 @@ using Blockcore.Connection;
 using Blockcore.Connection.Broadcasting;
 using Blockcore.Features.Wallet.Api.Controllers;
 using Blockcore.Features.Wallet.Api.Models;
+using Blockcore.Features.Wallet.Database;
 using Blockcore.Features.Wallet.Exceptions;
 using Blockcore.Features.Wallet.Interfaces;
 using Blockcore.Features.Wallet.Types;
