+++ src/Tests/Blockcore.IntegrationTests/Miners/ProofOfStakeMintCoinsTest.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Blockcore.Features.Miner.Interfaces;
 using Blockcore.Features.Miner.Staking;
 using Blockcore.Features.Wallet;
+using Blockcore.Features.Wallet.Database;
 using Blockcore.Features.Wallet.Types;
 using Blockcore.IntegrationTests.Common;
 using Blockcore.IntegrationTests.Common.EnvironmentMockUpHelpers;
